chore: move tests to top-level dir
